.blueprints/interactive_app/files/apps/{{blueprint_instance_name___kebab_case}}-cli/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
idna==3.4
inquirer==3.1.3
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.4
pyrsistent==0.19.3
python-editor==1.0.4
PyYAML==6.0
readchar==4.0.5
requests==2.28.2
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.8.2
